requirements-dev.txt
-r requirements.txt
pytest>=7.0.0
pytest-django>=4.5.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
